fix: refactor link handling in PDF utilities and improve type definitions
